Move launcher to its own module
