+++ modules/decima-app/build.gradle
@@ -1,0 +1,21 @@
+plugins {
+    id 'java'
+    id 'application'
+}
+
+application {
+    mainClass = 'com.shade.platform.Launcher'
+    mainModule = 'decima.main'
+    executableDir = ''
+    applicationDefaultJvmArgs = [
+            // Required for UIUtils#minimizePanel
+            '--add-opens', 'java.desktop/javax.swing.plaf.basic=ALL-UNNAMED'
+    ]
+}
+
+dependencies {
+    implementation project(':platform-model')
+    implementation project(':platform-ui')
+    implementation project(':decima-model')
+    implementation project(':decima-ui')
+}
