+++ modules/decima-app/src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module decima.main {
+module decima.app {
     requires decima.platform.ui;
     requires decima.platform;
     requires org.slf4j;
